week 7 progress add responsive
