+++ lib/feature/auth/login_page.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_rwid/feature/pages/dashboard.dart';
 import 'package:flutter_rwid/feature/auth/signup_page.dart';
+import 'package:flutter_rwid/feature/widgets/auth/login_form.dart';
 
 import 'package:flutter_rwid/utils/text_utils.dart';
 
@@ -30,177 +31,7 @@ class _LoginPageState extends State<LoginPage> {
           ),
         ),
         alignment: Alignment.center,
-        child: Form(
-          key: _formKey,
-          child: Container(
-            height: 400,
-            width: double.infinity,
-            margin: const EdgeInsets.symmetric(horizontal: 30),
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.white),
-              borderRadius: BorderRadius.circular(15),
-              color: Colors.black.withOpacity(0.4),
-            ),
-            child: ClipRRect(
-              borderRadius: BorderRadius.circular(20),
-              child: BackdropFilter(
-                filter: ImageFilter.blur(sigmaY: 5, sigmaX: 5),
-                child: Padding(
-                  padding: const EdgeInsets.all(25),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Spacer(),
-                      const Center(
-                        child: TextUtil(
-                          text: "Login",
-                          weight: true,
-                          size: 30,
-                        ),
-                      ),
-                      const Spacer(),
-                      const TextUtil(text: "Email"),
-                      Container(
-                        height: 35,
-                        decoration: const BoxDecoration(
-                          border: Border(
-                            bottom: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                        child: TextFormField(
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return "Please input your email";
-                            }
-                            if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$')
-                                .hasMatch(value)) {
-                              return "Please enter a valid email";
-                            }
-                            return null;
-                          },
-                          style: const TextStyle(color: Colors.white),
-                          decoration: const InputDecoration(
-                            suffixIcon: Icon(
-                              Icons.mail,
-                              color: Colors.white,
-                            ),
-                            fillColor: Colors.white,
-                            border: InputBorder.none,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      const TextUtil(text: "Password"),
-                      Container(
-                        height: 35,
-                        decoration: const BoxDecoration(
-                          border: Border(
-                            bottom: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                        child: TextFormField(
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return 'Please enter your password';
-                            }
-
-                            if (value.length < 6) {
-                              return "Password must be at least 6 characters";
-                            }
-                            return null;
-                          },
-                          obscureText: true,
-                          style: const TextStyle(color: Colors.white),
-                          decoration: const InputDecoration(
-                            suffixIcon: Icon(
-                              Icons.lock,
-                              color: Colors.white,
-                            ),
-                            fillColor: Colors.white,
-                            border: InputBorder.none,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      Row(
-                        children: [
-                          InkWell(
-                            onTap: () {
-                              setState(() {
-                                _rememberMe = !_rememberMe;
-                              });
-                            },
-                            child: Container(
-                                height: 15,
-                                width: 15,
-                                decoration: BoxDecoration(
-                                  color:
-                                      _rememberMe ? Colors.blue : Colors.white,
-                                  border: Border.all(color: Colors.white),
-                                ),
-                                child: _rememberMe
-                                    ? const Icon(Icons.check,
-                                        size: 12, color: Colors.white)
-                                    : null),
-                          ),
-                          const SizedBox(width: 10),
-                          const Expanded(
-                            child: TextUtil(
-                              text: "Remember Me",
-                              size: 12,
-                              weight: true,
-                            ),
-                          )
-                        ],
-                      ),
-                      const Spacer(),
-                      InkWell(
-                        onTap: () {
-                          if (_formKey.currentState!.validate()) {
-                            ScaffoldMessenger.of(context).showSnackBar(
-                              const SnackBar(content: Text('Success Login..')),
-                            );
-
-                            Navigator.of(context).push(MaterialPageRoute(
-                                builder: (context) => const Dashboard()));
-                          }
-                        },
-                        child: Container(
-                          height: 40,
-                          width: double.infinity,
-                          decoration: BoxDecoration(
-                            color: Colors.white,
-                            borderRadius: BorderRadius.circular(30),
-                          ),
-                          alignment: Alignment.center,
-                          child: const TextUtil(
-                            text: "Log In",
-                            color: Colors.black,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      Center(
-                        child: InkWell(
-                          onTap: () {
-                            Navigator.of(context).push(MaterialPageRoute(
-                                builder: (context) => const SignupPage()));
-                          },
-                          child: const TextUtil(
-                            text: "Don't have an account? REGISTER",
-                            size: 12,
-                            weight: true,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
+        child: LoginForm(),
       ),
     );
   }
