+++ lib/feature/auth/signup_page.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:flutter_rwid/feature/auth/login_page.dart';
 import 'package:flutter_rwid/feature/auth/signup_page.dart';
+import 'package:flutter_rwid/feature/widgets/auth/signup_form.dart';
 
 import 'package:flutter_rwid/utils/text_utils.dart';
 
@@ -14,18 +15,7 @@ class SignupPage extends StatefulWidget {
 }
 
 class _SignupPageState extends State<SignupPage> {
-  final _formKey = GlobalKey<FormState>();
-  final TextEditingController _passwordController = TextEditingController();
-  final TextEditingController _confirmPasswordController =
-      TextEditingController();
-
   @override
-  void dispose() {
-    _passwordController.dispose();
-    _confirmPasswordController.dispose();
-    super.dispose();
-  }
-
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -39,178 +29,7 @@ class _SignupPageState extends State<SignupPage> {
           ),
         ),
         alignment: Alignment.center,
-        child: Form(
-          key: _formKey,
-          child: Container(
-            height: 400,
-            width: double.infinity,
-            margin: const EdgeInsets.symmetric(horizontal: 30),
-            decoration: BoxDecoration(
-              border: Border.all(color: Colors.white),
-              borderRadius: BorderRadius.circular(15),
-              color: Colors.black.withOpacity(0.4),
-            ),
-            child: ClipRRect(
-              borderRadius: BorderRadius.circular(20),
-              child: BackdropFilter(
-                filter: ImageFilter.blur(sigmaY: 5, sigmaX: 5),
-                child: Padding(
-                  padding: const EdgeInsets.all(25),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Spacer(),
-                      const Center(
-                        child: TextUtil(
-                          text: "Sign Up",
-                          weight: true,
-                          size: 30,
-                        ),
-                      ),
-                      const Spacer(),
-                      const TextUtil(text: "Email"),
-                      Container(
-                        height: 35,
-                        decoration: const BoxDecoration(
-                          border: Border(
-                            bottom: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                        child: TextFormField(
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return "Please input your email";
-                            }
-                            if (!RegExp(r'^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$')
-                                .hasMatch(value)) {
-                              return "Please enter a valid email";
-                            }
-                            return null;
-                          },
-                          style: const TextStyle(color: Colors.white),
-                          decoration: const InputDecoration(
-                            suffixIcon: Icon(
-                              Icons.mail,
-                              color: Colors.white,
-                            ),
-                            fillColor: Colors.white,
-                            border: InputBorder.none,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      const TextUtil(text: "Password"),
-                      Container(
-                        height: 35,
-                        decoration: const BoxDecoration(
-                          border: Border(
-                            bottom: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                        child: TextFormField(
-                          controller: _passwordController,
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return "Please input your password";
-                            }
-                            if (value.length < 6) {
-                              return "Password must be at least 6 characters";
-                            }
-                            return null;
-                          },
-                          obscureText: true,
-                          style: const TextStyle(color: Colors.white),
-                          decoration: const InputDecoration(
-                            suffixIcon: Icon(
-                              Icons.lock,
-                              color: Colors.white,
-                            ),
-                            fillColor: Colors.white,
-                            border: InputBorder.none,
-                          ),
-                        ),
-                      ),
-                      const TextUtil(text: "Confirm Password"),
-                      Container(
-                        height: 35,
-                        decoration: const BoxDecoration(
-                          border: Border(
-                            bottom: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                        child: TextFormField(
-                          controller: _confirmPasswordController,
-                          validator: (value) {
-                            if (value == null || value.isEmpty) {
-                              return "Please confirm your password";
-                            }
-                            if (value != _passwordController.text) {
-                              return "Passwords do not match";
-                            }
-                            return null;
-                          },
-                          obscureText: true,
-                          style: const TextStyle(color: Colors.white),
-                          decoration: const InputDecoration(
-                            suffixIcon: Icon(
-                              Icons.lock,
-                              color: Colors.white,
-                            ),
-                            fillColor: Colors.white,
-                            border: InputBorder.none,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      InkWell(
-                        onTap: () {
-                          if (_formKey.currentState!.validate()) {
-                            if (_formKey.currentState!.validate()) {
-                              ScaffoldMessenger.of(context).showSnackBar(
-                                const SnackBar(
-                                    content: Text('Sign up successful!')),
-                              );
-                              Navigator.of(context).push(MaterialPageRoute(
-                                  builder: (context) => LoginPage()));
-                            }
-                          }
-                        },
-                        child: Container(
-                          height: 40,
-                          width: double.infinity,
-                          decoration: BoxDecoration(
-                            color: Colors.white,
-                            borderRadius: BorderRadius.circular(30),
-                          ),
-                          alignment: Alignment.center,
-                          child: const TextUtil(
-                            text: "Sign Up",
-                            color: Colors.black,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                      Center(
-                        child: InkWell(
-                          onTap: () {
-                            Navigator.of(context).push(MaterialPageRoute(
-                                builder: (context) => LoginPage()));
-                          },
-                          child: const TextUtil(
-                            text: "Have an Account? Login here!",
-                            size: 12,
-                            weight: true,
-                          ),
-                        ),
-                      ),
-                      const Spacer(),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
+        child: SignupForm(),
       ),
     );
   }
